Split app.js phase 2a: extract override-challenge, update-banner, theme modules
Mechanical moves; temporary imports from './app.js' for functions that will
move to their final module in later phases. In-app tests pass (132/132).
